// src/app/gateway-template/model/gateway-template-io-tag.ts
export class GatewayTemplateIOTag {

    id: number;
    name: string;
    nodeId: number;
    tagKeyName: string;
    description: string;
    tagType: string;
    dataTypeId: any;
    dataProtocolStandardTagId: string;
    engUnitId: string;
    status;
    created_by: number;
    updated_by: number;
    tagIOMode: any;
    dataProtocolStandardTag:number
}